京喜工厂可生产
